Update associated with 1 May 2020 memo
different approach to dummy variables that did so for most, but not all, of the categorical variables used. The exception was Multi.Aircraft, which was already binary. The reset had three or more category levels. for dummyVars, Full rank option was false (creating linear dependent dummy variables) due to the expectation that one or more of the dummy variables for each feature would be of low or no variance. Therefore, if at least one is of low or no rank, no linear dependencies If that is not the case, at least one dummy variable will be removed  from each categorical feature to prevent linear dependencies.