Included alternative scaling
The key predictions were one that took 25% of the earlier data to predict the Severity.Loss of the newer records. Adde a twist to the end. Assumed that new records would be tested singly and not as a group, so numeric scaling transformed individual numeric values based on the transformation used on the training set. For example, for a normalized training feature, subtract that feature's mean from the test record feature, and divide by the training features standard deviation.